+++ web-wallet/src/lib/components/index.js
@@ -4,6 +4,7 @@ export { default as AppAnchor } from "./AppAnchor/AppAnchor.svelte";
 export { default as AppAnchorButton } from "./AppAnchorButton/AppAnchorButton.svelte";
 export { default as AppImage } from "./AppImage/AppImage.svelte";
 export { default as ApproveMigration } from "./ApproveMigration/ApproveMigration.svelte";
+export { default as Banner } from "./Banner/Banner.svelte";
 export { default as Balance } from "./Balance/Balance.svelte";
 export { default as ContractOperations } from "./ContractOperations/ContractOperations.svelte";
 export { default as ContractStatusesList } from "./ContractStatusesList/ContractStatusesList.svelte";