+++ web-wallet/src/routes/(welcome)/setup/create/MnemonicValidate.svelte
@@ -1,8 +1,8 @@
 <script>
-  import { Icon, Mnemonic } from "$lib/dusk/components";
+  import { Mnemonic } from "$lib/dusk/components";
   import { arraysEqual, shuffleArray } from "$lib/dusk/array";
-  import { mdiAlertOutline } from "@mdi/js";
   import { IconHeadingCard } from "$lib/containers/Cards";
+  import { Banner } from "$lib/components";
 
   /** @type {boolean} */
   export let isValid = false;
@@ -19,23 +19,18 @@
   $: isValid = arraysEqual(enteredMnemonicPhrase, mnemonicPhrase);
 </script>
 
-<IconHeadingCard heading="Verification">
-  <p class="mnemonic-notice">Ensure you have backed up the Mnemonic phrase.</p>
+<IconHeadingCard heading="Verification" gap="medium">
+  <p>Ensure you have backed up the Mnemonic phrase.</p>
   <Mnemonic
     bind:enteredMnemonicPhrase
     mnemonicPhrase={shuffleArray(mnemonicPhrase)}
     type="validate"
   />
   {#if filteredMnemonic.size === 12 && !isValid}
-    <div class="notice notice--error">
-      <Icon path={mdiAlertOutline} size="large" />
-      <p>Mnemonic does not match.</p>
-    </div>
+    <Banner title="Mnemonic does not match." variant="error">
+      <p>
+        Please ensure you have entered the mnemonic phrase in the correct order.
+      </p>
+    </Banner>
   {/if}
 </IconHeadingCard>
-
-<style>
-  .mnemonic-notice {
-    margin-bottom: var(--small-gap);
-  }
-</style>