+++ web-wallet/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add validation for public account ("Send" flow) [#2176]
 - Added allocation (shield/unshield) page and UI [#2196]
 - Add validation for "Use Max" button on Send / Stake flows [#2310]
+- Add Banner component [#2696]
 
 ### Changed
 
@@ -278,6 +279,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2436]: https://github.com/dusk-network/rusk/issues/2436
 [#2498]: https://github.com/dusk-network/rusk/issues/2498
 [#2532]: https://github.com/dusk-network/rusk/issues/2532
+[#2696]: https://github.com/dusk-network/rusk/issues/2696
 
 <!-- VERSIONS -->